pickle and hdf5 functions moved to io package
