+++ src/rsatoolbox/data/dataset.py
@@ -20,8 +20,8 @@ from rsatoolbox.util.descriptor_utils import format_descriptor
 from rsatoolbox.util.descriptor_utils import parse_input_descriptor
 from rsatoolbox.util.descriptor_utils import append_obs_descriptors
 from rsatoolbox.util.descriptor_utils import desc_eq
-from rsatoolbox.util.file_io import read_dict_hdf5
-from rsatoolbox.util.file_io import read_dict_pkl
+from rsatoolbox.io.hdf5 import read_dict_hdf5
+from rsatoolbox.io.pkl import read_dict_pkl
 from rsatoolbox.data.base import DatasetBase
