+++ tests/test_io_hdf5.py
@@ -5,11 +5,11 @@ from importlib.metadata import version
 
 class Hdf5IOTests(TestCase):
 
-    @patch('rsatoolbox.util.file_io.File')
+    @patch('rsatoolbox.io.hdf5.File')
     def test_write_dict_hdf5_version(self, h5pyFile):
         """Check version tag matches current version
         """
-        from rsatoolbox.util.file_io import write_dict_hdf5
+        from rsatoolbox.io.hdf5 import write_dict_hdf5
         h5pyFile().attrs = dict()   
         write_dict_hdf5('a file path', dict())
         self.assertEqual(
